+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tokiforge/angular
 
+## 2.0.1
+
+### Patch Changes
+
+- Version bump; @tokiforge/core@^2.0.1.
+
 ## 2.0.0
 
 ### Major Changes