+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tokiforge/svelte
 
+## 2.0.1
+
+### Patch Changes
+
+- Version bump; dependency range ^2.0.1.
+
 ## 2.0.0
 
 ### Major Changes